requirements.txt
uvicorn==0.30.1
fastapi==0.111.0
geojson==3.1.0
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==10.3.0
rasterio==1.3.10
Shapely==2.0.4
ultralytics==8.2.32

